studioManager module
move codes from common space to new studioManager module
