+++ src/src/modules/studioManager/index.js
@@ -1,0 +1,21 @@
+import StudioManager from "./StudioManager"
+
+var IDE = window.IDE;
+
+export default {
+
+	activate(loaded) {
+
+        if (typeof studio !== 'undefined') {
+            this.init();
+        }
+
+        loaded();
+		
+	},
+
+    init() {
+        var studioManager    = new StudioManager();
+        IDE.studioManager    = studioManager;
+    }
+}
